Refactor: Make class public. DefaultPasswordEncoderAuthenticationManagerBuilder, LazyPasswordEncoder
